Removed GraphMetadata API point
Removed most of GraphMetadata functionality and moved it to Graph
(transitModes) or RouterInfo (center).

And renamed Graphmetadata to GraphEnvelope.
