+++ src/main/java/org/opentripplanner/graph_builder/module/GtfsModule.java
@@ -149,7 +149,6 @@ public class GtfsModule implements GraphBuilderModule {
 
         graph.hasTransit = true;
         graph.calculateTransitCenter();
-        graph.getMetadata().addCenter(graph.getCenter());
 
     }
